feat(gopro): use cori and iori instead of gyro when available
On GoPro Hero 8, 9 and 10 there are CORI and IORI gpmf data which is
exact camera orientation and image orientation respectively per frame.
Uses that intead of gyro data if it is available.

This IORI is the image orientation relative to camera body, which means
we can now stabilize videos recorded with hypersmooth.

TODO: Make this work in the GUI. Only works via CLI right now.

fixes #51